+++ mod/lesson/locallib.php
@@ -3499,6 +3499,7 @@ class lesson extends lesson_base {
             'displayscorewithessays' => false,
             'displayscorewithoutessays' => false,
             'yourcurrentgradeisoutof' => false,
+            'yourcurrentgradeis' => false,
             'eolstudentoutoftimenoanswers' => false,
             'welldone' => false,
             'progressbar' => false,
@@ -3589,12 +3590,6 @@ class lesson extends lesson_base {
                     } else {
                         $data->displayscorewithoutessays = $a;
                     }
-                    if ($this->properties->grade != GRADE_TYPE_NONE) {
-                        $a = new stdClass;
-                        $a->grade = format_float($gradeinfo->grade * $this->properties->grade / 100, 1);
-                        $a->total = $this->properties->grade;
-                        $data->yourcurrentgradeisoutof = $a;
-                    }
 
                     $grade = new stdClass();
                     $grade->lessonid = $this->properties->id;
@@ -3612,6 +3607,25 @@ class lesson extends lesson_base {
                     } else {
                         $newgradeid = $DB->insert_record("lesson_grades", $grade);
                     }
+
+                    // Update central gradebook.
+                    lesson_update_grades($this, $USER->id);
+
+                    // Print grade (grade type Point).
+                    if ($this->properties->grade > 0) {
+                        $a = new stdClass;
+                        $a->grade = format_float($gradeinfo->grade * $this->properties->grade / 100, 1);
+                        $a->total = $this->properties->grade;
+                        $data->yourcurrentgradeisoutof = $a;
+                    }
+
+                    // Print grade (grade type Scale).
+                    if ($this->properties->grade < 0) {
+                        // Grade type is Scale.
+                        $grades = grade_get_grades($course->id, 'mod', 'lesson', $cm->instance, $USER->id);
+                        $grade = reset($grades->items[0]->grades);
+                        $data->yourcurrentgradeis = $grade->str_grade;
+                    }
                 } else {
                     if ($this->properties->timelimit) {
                         if ($outoftime == 'normal') {
@@ -3622,14 +3636,15 @@ class lesson extends lesson_base {
                             $grade->completed = time();
                             $newgradeid = $DB->insert_record("lesson_grades", $grade);
                             $data->eolstudentoutoftimenoanswers = true;
+
+                            // Update central gradebook.
+                            lesson_update_grades($this, $USER->id);
                         }
                     } else {
                         $data->welldone = true;
                     }
                 }
 
-                // Update central gradebook.
-                lesson_update_grades($this, $USER->id);
                 $data->progresscompleted = $progresscompleted;
             }
         } else {
@@ -4224,15 +4239,20 @@ abstract class lesson_page extends lesson_base {
                 }
                 // "number of attempts remaining" message if $this->lesson->maxattempts > 1
                 // displaying of message(s) is at the end of page for more ergonomic display
-                if (!$result->correctanswer && ($result->newpageid == 0)) {
+                // If we are showing the number of remaining attempts, we need to show it regardless of what the next
+                // jump to page is.
+                if (!$result->correctanswer) {
                     // Retrieve the number of attempts left counter for displaying at bottom of feedback page.
-                    if (!empty($this->lesson->maxattempts) && $nattempts >= $this->lesson->maxattempts) {
+                    if ($result->newpageid == 0 && !empty($this->lesson->maxattempts) && $nattempts >= $this->lesson->maxattempts) {
                         if ($this->lesson->maxattempts > 1) { // don't bother with message if only one attempt
                             $result->maxattemptsreached = true;
                         }
                         $result->newpageid = LESSON_NEXTPAGE;
                     } else if ($this->lesson->maxattempts > 1) { // don't bother with message if only one attempt
                         $result->attemptsremaining = $this->lesson->maxattempts - $nattempts;
+                        if ($result->attemptsremaining == 0) {
+                            $result->maxattemptsreached = true;
+                        }
                     }
                 }
             }