+++ mod/minilesson/amd/src/dialogplayer.js
@@ -1,0 +1,21 @@
+define(['jquery','core/log'], function($,log) {
+    "use strict"; // jshint ;_;
+
+/*
+This file contains class and ID definitions.
+ */
+
+    log.debug('MiniLesson Conversation player: initialising');
+
+    return{
+        //pass in config, amd set up table
+        init: function(props){
+            //pick up opts from html
+            var that=this;
+
+        },
+
+
+
+    };//end of return value
+});